+++ docs/README.md
@@ -47,7 +47,7 @@
 - [x] 사용성 고민: 자동 포커스 기능을 통해 사용자가 프로그램 시작되면 가장 먼저해야할일 암시
 - [x] alert는 void형이기 때문에, 삼항연산자 지양 
 - [x] hidden없애고 id로 하기 
-- [ ] 상수 (하드코딩 x)
+- [x] 상수 (하드코딩 x)
 - [ ] 파일 정리해보기 (여러개인것 합치기)
 - [ ] 만약 재시작한다면 init? (드림코딩 참고)
