--- src/js/views/disableCarInput.js
@@ -1,8 +1,0 @@
-import $ from '../utils/dom.js';
-
-const disableCarInput = () => {
-  $('#car-names-input').disabled = 'true';
-  $('#car-names-submit').disabled = 'true';
-};
-
-export default disableCarInput;
